+++ CHANGES.rst
@@ -50,6 +50,7 @@ v3.0.0 (xxxx-xx-xx)
   - ``chars``
   - ``count_substr``
   - ``ends_with``
+  - ``quote``, ``surround``
 
 - Make ``capitalize`` accept ``lower_rest`` argument to determine whether to convert the rest of the string to lower case or not. Defaults to ``True``.
 - Make ``splice`` work with strings.